% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplicial_complexes.R
\name{alpha_complex}
\alias{alpha_complex}
\title{Alpha Complex Class}
\usage{
alpha_complex(points, precision = "safe")
}
\arguments{
\item{points}{Either a nxd matrix or a length-n list of d-dimensional vectors
or a file with extension \code{.off}.}

\item{precision}{A string specifying the alpha complex precision. Can be one
of \code{"fast"}, \code{"safe"} or \code{"exact"}. Defaults to \code{"safe"}.}
}
\description{
AlphaComplex is a simplicial complex constructed from the finite cells of a
Delaunay Triangulation.
}
\details{
The filtration value of each simplex is computed as the square of the
circumradius of the simplex if the circumsphere is empty (the simplex is then
said to be Gabriel), and as the minimum of the filtration values of the
codimension 1 cofaces that make it not Gabriel otherwise. All simplices that
have a filtration value strictly greater than a given alpha squared value are
not inserted into the complex.
}
\examples{
n <- 20
X_list <- replicate(n, runif(2), simplify = FALSE)
X_matrix <- Reduce(rbind, X_list, init = numeric())
ac_matrix <- alpha_complex(points = X_matrix)
st_matrix <- ac_matrix$create_simplex_tree()
dgm_matrix <- st_matrix$persistence()
plot_persistence_diagram(dgm_matrix)
ac_list <- alpha_complex(points = X_list)
st_list <- ac_list$create_simplex_tree()
dgm_list <- st_list$persistence()
plot_persistence_diagram(dgm_list)
}
