% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perslay.R
\name{BasePerslayWeightClass}
\alias{BasePerslayWeightClass}
\title{Perslay: Base Perslay Weight Class}
\description{
This is a base class for computing a differentiable weight
function for persistence diagram points.
}
\keyword{internal}
\section{Super classes}{
\code{rgudhi::PythonClass} -> \code{\link[rgudhi:PerslayBaseClass]{rgudhi::PerslayBaseClass}} -> \code{BasePerslayWeightClass}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BasePerslayWeightClass-call}{\code{BasePerslayWeightClass$call()}}
\item \href{#method-BasePerslayWeightClass-clone}{\code{BasePerslayWeightClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="get_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-get_python_class'><code>rgudhi::PythonClass$get_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="set_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-set_python_class'><code>rgudhi::PythonClass$set_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PerslayBaseClass" data-id="build"><a href='../../rgudhi/html/PerslayBaseClass.html#method-PerslayBaseClass-build'><code>rgudhi::PerslayBaseClass$build()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasePerslayWeightClass-call"></a>}}
\if{latex}{\out{\hypertarget{method-BasePerslayWeightClass-call}{}}}
\subsection{Method \code{call()}}{
Apply the layer to an input, or list of inputs, and return
the result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasePerslayWeightClass$call(diagrams)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{diagrams}}{A ragged tensor of shape \eqn{n \times \mathrm{None} \times 2} specifying
a sample of \eqn{n} persistence diagrams. The second dimension is ragged
because the number of points in each diagram may vary.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ragged tensor storing the weights of the points in the \eqn{n}
persistence diagrams. The second dimension is ragged since persistence
diagrams can have different numbers of points.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasePerslayWeightClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BasePerslayWeightClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasePerslayWeightClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
