% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn-cluster.R
\name{ClusteringAlgorithm}
\alias{ClusteringAlgorithm}
\title{Base Class for Clustering Algorithms}
\description{
This is the base class for all the clustering algorithms in the
\href{https://scikit-learn.org/stable/modules/classes.html#module-sklearn.cluster}{\strong{sklearn.cluster}}
module. The child classes are intended to be used within some GUDHI classes
such as \link{Atol}.
}
\keyword{internal}
\section{Super class}{
\code{rgudhi::PythonClass} -> \code{ClusteringAlgorithm}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClusteringAlgorithm-clone}{\code{ClusteringAlgorithm$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="get_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-get_python_class'><code>rgudhi::PythonClass$get_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="set_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-set_python_class'><code>rgudhi::PythonClass$set_python_class()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringAlgorithm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringAlgorithm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringAlgorithm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
