% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perslay.R
\name{GaussianMixturePerslayWeight}
\alias{GaussianMixturePerslayWeight}
\title{Perslay: Gaussian Mixture Perslay Weight}
\description{
This is a class for computing a differentiable weight function
for persistence diagram points. This function is defined from a mixture of
Gaussian functions.
}
\examples{
\dontshow{if (reticulate::py_module_available("gudhi")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
diagrams <- as_persistence_diagram_sample(list(
  as_persistence_diagram(tibble::tibble(
    birth = c(0, 1, 3, 6),
    death = c(4, 2, 8, 8)
  ))
))
ds <- DiagramScaler$new(
  use = TRUE,
  scalers = list(list(c(0, 1), MinMaxScaler$new()))
)
dgms <- diagrams |>
  ds$fit_transform() |>
  to_ragged_tensor()

weight <- GaussianMixturePerslayWeight$new(
  gaussians = matrix(c(0, 0, 1, 1), nrow = 4, ncol = 1)
)
weight$call(dgms)
\dontshow{\}) # examplesIf}
}
\author{
Mathieu Carrière, Martin Royer, Gard Spreemann, Wojciech Reise
}
\section{Super classes}{
\code{rgudhi::PythonClass} -> \code{\link[rgudhi:PerslayBaseClass]{rgudhi::PerslayBaseClass}} -> \code{\link[rgudhi:BasePerslayWeightClass]{rgudhi::BasePerslayWeightClass}} -> \code{GaussianMixturePerslayWeight}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GaussianMixturePerslayWeight-new}{\code{GaussianMixturePerslayWeight$new()}}
\item \href{#method-GaussianMixturePerslayWeight-clone}{\code{GaussianMixturePerslayWeight$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="get_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-get_python_class'><code>rgudhi::PythonClass$get_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="set_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-set_python_class'><code>rgudhi::PythonClass$set_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PerslayBaseClass" data-id="build"><a href='../../rgudhi/html/PerslayBaseClass.html#method-PerslayBaseClass-build'><code>rgudhi::PerslayBaseClass$build()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="BasePerslayWeightClass" data-id="call"><a href='../../rgudhi/html/BasePerslayWeightClass.html#method-BasePerslayWeightClass-call'><code>rgudhi::BasePerslayWeightClass$call()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianMixturePerslayWeight-new"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianMixturePerslayWeight-new}{}}}
\subsection{Method \code{new()}}{
The \code{\link{GaussianMixturePerslayWeight}} constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianMixturePerslayWeight$new(gaussians, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gaussians}}{A numeric matrix of size \eqn{4 \times n} specifying a
two-dimensional Gaussian distribution for each of the \eqn{n} diagrams.
Each column must store the Gaussian parameters in the following order:
\eqn{\mu_x, \mu_y, \sigma_x, \sigma_y}.}

\item{\code{...}}{A named list providing extra arguments for compatibility with
the TensorFlow API. Not used here.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{\link{GaussianMixturePerslayWeight}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianMixturePerslayWeight-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianMixturePerslayWeight-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianMixturePerslayWeight$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
