% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perslay.R
\name{GaussianPerslayPhi}
\alias{GaussianPerslayPhi}
\title{Perslay: Gaussian Perslay Phi Class}
\description{
This is a class for computing a transformation function for
persistence diagram points. This function turns persistence diagram points
into 2D Gaussian functions centered on the points, that are then evaluated
on a regular 2D grid.
}
\examples{
\dontshow{if (reticulate::py_module_available("gudhi")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
diagrams <- as_persistence_diagram_sample(list(
  as_persistence_diagram(tibble::tibble(
    birth = c(0, 1, 3, 6),
    death = c(4, 2, 8, 8)
  ))
))
ds <- DiagramScaler$new(
  use = TRUE,
  scalers = list(list(c(0, 1), MinMaxScaler$new()))
)
dgms <- diagrams |>
  ds$fit_transform() |>
  to_ragged_tensor()

phi <- GaussianPerslayPhi$new(
  image_size = c(5, 5),
  image_bnds = rbind(c(-.5, 1.5), c(-.5, 1.5)),
  variance = .1
)
phi$call(dgms)
\dontshow{\}) # examplesIf}
}
\author{
Mathieu Carrière, Martin Royer, Gard Spreemann, Wojciech Reise
}
\section{Super classes}{
\code{rgudhi::PythonClass} -> \code{\link[rgudhi:PerslayBaseClass]{rgudhi::PerslayBaseClass}} -> \code{\link[rgudhi:BasePerslayPhiClass]{rgudhi::BasePerslayPhiClass}} -> \code{GaussianPerslayPhi}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GaussianPerslayPhi-new}{\code{GaussianPerslayPhi$new()}}
\item \href{#method-GaussianPerslayPhi-clone}{\code{GaussianPerslayPhi$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="get_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-get_python_class'><code>rgudhi::PythonClass$get_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="set_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-set_python_class'><code>rgudhi::PythonClass$set_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PerslayBaseClass" data-id="build"><a href='../../rgudhi/html/PerslayBaseClass.html#method-PerslayBaseClass-build'><code>rgudhi::PerslayBaseClass$build()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="BasePerslayPhiClass" data-id="call"><a href='../../rgudhi/html/BasePerslayPhiClass.html#method-BasePerslayPhiClass-call'><code>rgudhi::BasePerslayPhiClass$call()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianPerslayPhi-new"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianPerslayPhi-new}{}}}
\subsection{Method \code{new()}}{
The \code{\link{GaussianPerslayPhi}} constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianPerslayPhi$new(image_size, image_bnds, variance, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image_size}}{A length-2 integer vector specifying the size of the 2D
grid on which the phi function must be evaluated.}

\item{\code{image_bnds}}{An integer matrix of shape \eqn{2 \times 2} specifying
the boundaries of the grid on which the phi function must be evaluated.
It must be of the form \eqn{\begin{bmatrix} x_{min} & x_{max} \\
  y_{min} & y_{max} \end{bmatrix}}.}

\item{\code{variance}}{A numeric value specifying the variance of the Gaussian
functions.}

\item{\code{...}}{A named list providing extra arguments for compatibility with the TensorFlow
API. Not used here.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{\link{GaussianPerslayPhi}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianPerslayPhi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianPerslayPhi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianPerslayPhi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
