% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perslay.R
\name{GridPerslayWeight}
\alias{GridPerslayWeight}
\title{Perslay: Grid Perslay Weight}
\description{
This is a class for computing a differentiable weight function
for persistence diagram points. This function is defined from an array that
contains its values on a 2D grid.
}
\examples{
\dontshow{if (reticulate::py_module_available("gudhi")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
diagrams <- as_persistence_diagram_sample(list(
  as_persistence_diagram(tibble::tibble(
    birth = c(0, 1, 3, 6),
    death = c(4, 2, 8, 8)
  ))
))
ds <- DiagramScaler$new(
  use = TRUE,
  scalers = list(list(c(0, 1), MinMaxScaler$new()))
)
dgms <- diagrams |>
  ds$fit_transform() |>
  to_ragged_tensor()

weight <- GridPerslayWeight$new(
  grid = matrix(c(1:13, 12:1), 5, 5),
  grid_bnds = rbind(c(-.5, 1.5), c(-.5, 1.5))
)
weight$call(dgms)
\dontshow{\}) # examplesIf}
}
\author{
Mathieu Carrière, Martin Royer, Gard Spreemann, Wojciech Reise
}
\section{Super classes}{
\code{rgudhi::PythonClass} -> \code{\link[rgudhi:PerslayBaseClass]{rgudhi::PerslayBaseClass}} -> \code{\link[rgudhi:BasePerslayWeightClass]{rgudhi::BasePerslayWeightClass}} -> \code{GridPerslayWeight}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GridPerslayWeight-new}{\code{GridPerslayWeight$new()}}
\item \href{#method-GridPerslayWeight-clone}{\code{GridPerslayWeight$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="get_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-get_python_class'><code>rgudhi::PythonClass$get_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="set_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-set_python_class'><code>rgudhi::PythonClass$set_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PerslayBaseClass" data-id="build"><a href='../../rgudhi/html/PerslayBaseClass.html#method-PerslayBaseClass-build'><code>rgudhi::PerslayBaseClass$build()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="BasePerslayWeightClass" data-id="call"><a href='../../rgudhi/html/BasePerslayWeightClass.html#method-BasePerslayWeightClass-call'><code>rgudhi::BasePerslayWeightClass$call()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridPerslayWeight-new"></a>}}
\if{latex}{\out{\hypertarget{method-GridPerslayWeight-new}{}}}
\subsection{Method \code{new()}}{
The \code{\link{GridPerslayWeight}} constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridPerslayWeight$new(grid, grid_bnds, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{grid}}{A numeric matrix of shape \eqn{n \times n} specifying the
grid of values.}

\item{\code{grid_bnds}}{A numeric matrix of shape \eqn{2 \times 2} specifying
the boundaries of the grid. It should be of the form
\eqn{\begin{bmatrix} x_{min} & x_{max} \\ y_{min} & y_{max}
  \end{bmatrix}}.}

\item{\code{...}}{A named list providing extra arguments for compatibility with
the TensorFlow API. Not used here.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{\link{GridPerslayWeight}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridPerslayWeight-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GridPerslayWeight-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridPerslayWeight$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
