% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence-diagram.R
\name{persistence_diagram}
\alias{persistence_diagram}
\alias{as_persistence_diagram}
\alias{is_persistence_diagram}
\title{Persistence Diagram}
\usage{
as_persistence_diagram(x)

is_persistence_diagram(x)
}
\arguments{
\item{x}{An object coercible into a \link{persistence_diagram} object.}
}
\value{
An object of class \link{persistence_diagram}.
}
\description{
A collection of function to manipulate a persistence diagram as an object of
class \link{persistence_diagram}. A \link{persistence_diagram} is a \link[tibble:tibble]{tibble::tibble}
with a \code{birth} variable and at least one of \code{death} or \code{lifetime} variables.
}
