% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence_diagram_sample.R
\name{persistence_diagram_sample}
\alias{persistence_diagram_sample}
\alias{as_persistence_diagram_sample}
\alias{is_persistence_diagram_sample}
\title{Persistence Diagram Sample}
\usage{
as_persistence_diagram_sample(x)

is_persistence_diagram_sample(x)
}
\arguments{
\item{x}{An object coercible into a \link{persistence_diagram_sample} object.}
}
\value{
An object of class \link{persistence_diagram_sample}.
}
\description{
A collection of function to manipulate a persistence diagram sample as an
object of class \link{persistence_diagram_sample}. A \link{persistence_diagram_sample}
is a list of objects of class \link{persistence_diagram}.
}
