% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplicial_complexes.R
\name{plot_persistence_diagram}
\alias{plot_persistence_diagram}
\title{Plot persistence diagrams}
\usage{
plot_persistence_diagram(diagram)
}
\arguments{
\item{diagram}{A persistent diagram.}
}
\description{
Plot persistence diagrams
}
\examples{
n <- 10
X <- replicate(n, runif(2), simplify = FALSE)
ac <- alpha_complex(points = X)
st <- ac$create_simplex_tree()
dgm <- st$persistence()
plot_persistence_diagram(dgm)
}
