% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perslay.R
\name{PowerPerslayWeight}
\alias{PowerPerslayWeight}
\title{Perslay: Power Perslay Weight}
\description{
This is a class for computing a differentiable weight function
for persistence diagram points. This function is defined as a constant
multiplied by the distance to the diagonal of the persistence diagram point
raised to some power.
}
\examples{
\dontshow{if (reticulate::py_module_available("gudhi")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
diagrams <- as_persistence_diagram_sample(list(
  as_persistence_diagram(tibble::tibble(
    birth = c(0, 1, 3, 6),
    death = c(4, 2, 8, 8)
  ))
))
ds <- DiagramScaler$new(
  use = TRUE,
  scalers = list(list(c(0, 1), MinMaxScaler$new()))
)
dgms <- diagrams |>
  ds$fit_transform() |>
  to_ragged_tensor()

weight <- PowerPerslayWeight$new(
  constant = 1,
  power = 0
)
weight$call(dgms)
\dontshow{\}) # examplesIf}
}
\author{
Mathieu Carrière, Martin Royer, Gard Spreemann, Wojciech Reise
}
\section{Super classes}{
\code{rgudhi::PythonClass} -> \code{\link[rgudhi:PerslayBaseClass]{rgudhi::PerslayBaseClass}} -> \code{PowerPerslayWeight}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PowerPerslayWeight-new}{\code{PowerPerslayWeight$new()}}
\item \href{#method-PowerPerslayWeight-clone}{\code{PowerPerslayWeight$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="get_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-get_python_class'><code>rgudhi::PythonClass$get_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="set_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-set_python_class'><code>rgudhi::PythonClass$set_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PerslayBaseClass" data-id="build"><a href='../../rgudhi/html/PerslayBaseClass.html#method-PerslayBaseClass-build'><code>rgudhi::PerslayBaseClass$build()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PerslayBaseClass" data-id="call"><a href='../../rgudhi/html/PerslayBaseClass.html#method-PerslayBaseClass-call'><code>rgudhi::PerslayBaseClass$call()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PowerPerslayWeight-new"></a>}}
\if{latex}{\out{\hypertarget{method-PowerPerslayWeight-new}{}}}
\subsection{Method \code{new()}}{
The \code{\link{PowerPerslayWeight}} constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PowerPerslayWeight$new(constant, power, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{constant}}{A numeric value specifying the constant of the
transformation.}

\item{\code{power}}{A numeric value specifying the power of the transformation
which will be applied to the distance to the diagonal.}

\item{\code{...}}{A named list providing extra arguments for compatibility with
the TensorFlow API. Not used here.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{\link{PowerPerslayWeight}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PowerPerslayWeight-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PowerPerslayWeight-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PowerPerslayWeight$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
