% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representations.R
\name{RepresentationBaseClass}
\alias{RepresentationBaseClass}
\title{Representation Module - Abstract Base Class}
\description{
Representation Module - Abstract Base Class

Representation Module - Abstract Base Class
}
\author{
Mathieu Carrière
}
\keyword{internal}
\section{Super classes}{
\code{rgudhi::PythonClass} -> \code{rgudhi::SKLearnClass} -> \code{RepresentationBaseClass}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RepresentationBaseClass-get_value}{\code{RepresentationBaseClass$get_value()}}
\item \href{#method-RepresentationBaseClass-fit}{\code{RepresentationBaseClass$fit()}}
\item \href{#method-RepresentationBaseClass-transform}{\code{RepresentationBaseClass$transform()}}
\item \href{#method-RepresentationBaseClass-fit_transform}{\code{RepresentationBaseClass$fit_transform()}}
\item \href{#method-RepresentationBaseClass-clone}{\code{RepresentationBaseClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="get_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-get_python_class'><code>rgudhi::PythonClass$get_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="set_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-set_python_class'><code>rgudhi::PythonClass$set_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="SKLearnClass" data-id="get_params"><a href='../../rgudhi/html/SKLearnClass.html#method-SKLearnClass-get_params'><code>rgudhi::SKLearnClass$get_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="SKLearnClass" data-id="set_params"><a href='../../rgudhi/html/SKLearnClass.html#method-SKLearnClass-set_params'><code>rgudhi::SKLearnClass$set_params()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentationBaseClass-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentationBaseClass-get_value}{}}}
\subsection{Method \code{get_value()}}{
Applies the class on a single persistence diagram and output
the result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentationBaseClass$get_value(diag)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{diag}}{A \link[tibble:tibble]{tibble::tibble} with two variables \code{birth} and \code{death}
specifying a persistence diagram.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentationBaseClass-fit"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentationBaseClass-fit}{}}}
\subsection{Method \code{fit()}}{
Fits the class on a list of persistence diagrams. It might
do nothing in some cases but is useful when the class is included in a
\strong{scikit-learn} pipeline.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentationBaseClass$fit(X, y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A list of \link[tibble:tibble]{tibble::tibble}s specifying a sample of persistence
diagrams.}

\item{\code{y}}{An integer vector specifying persistence diagram labels (unused for
now).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentationBaseClass-transform"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentationBaseClass-transform}{}}}
\subsection{Method \code{transform()}}{
Applies the class on the persistence diagrams.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentationBaseClass$transform(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A list of \link[tibble:tibble]{tibble::tibble}s specifying a sample of persistence
diagrams.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentationBaseClass-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentationBaseClass-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
Applies sequentially the \verb{$fit()} and \verb{$transform()} methods
on the persistence diagrams.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentationBaseClass$fit_transform(X, y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A list of \link[tibble:tibble]{tibble::tibble}s specifying a sample of persistence
diagrams.}

\item{\code{y}}{An integer vector specifying persistence diagram labels (unused for
now).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentationBaseClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentationBaseClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentationBaseClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
