% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-preprocessing.R
\name{RepresentationPreprocessingBaseClass}
\alias{RepresentationPreprocessingBaseClass}
\title{Abstract Class for Representation Preprocessing Module}
\description{
Abstract Class for Representation Preprocessing Module

Abstract Class for Representation Preprocessing Module
}
\author{
Mathieu Carrière
}
\keyword{internal}
\section{Super classes}{
\code{rgudhi::PythonClass} -> \code{rgudhi::SKLearnClass} -> \code{RepresentationPreprocessingBaseClass}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RepresentationPreprocessingBaseClass-get_value}{\code{RepresentationPreprocessingBaseClass$get_value()}}
\item \href{#method-RepresentationPreprocessingBaseClass-fit}{\code{RepresentationPreprocessingBaseClass$fit()}}
\item \href{#method-RepresentationPreprocessingBaseClass-transform}{\code{RepresentationPreprocessingBaseClass$transform()}}
\item \href{#method-RepresentationPreprocessingBaseClass-clone}{\code{RepresentationPreprocessingBaseClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="get_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-get_python_class'><code>rgudhi::PythonClass$get_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="set_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-set_python_class'><code>rgudhi::PythonClass$set_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="SKLearnClass" data-id="fit_transform"><a href='../../rgudhi/html/SKLearnClass.html#method-SKLearnClass-fit_transform'><code>rgudhi::SKLearnClass$fit_transform()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="SKLearnClass" data-id="get_params"><a href='../../rgudhi/html/SKLearnClass.html#method-SKLearnClass-get_params'><code>rgudhi::SKLearnClass$get_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="SKLearnClass" data-id="set_params"><a href='../../rgudhi/html/SKLearnClass.html#method-SKLearnClass-set_params'><code>rgudhi::SKLearnClass$set_params()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentationPreprocessingBaseClass-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentationPreprocessingBaseClass-get_value}{}}}
\subsection{Method \code{get_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentationPreprocessingBaseClass$get_value(diag)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{diag}}{A \link[tibble:tibble]{tibble::tibble} with two variables \code{birth} and \code{death}
specifying a persistence diagram.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentationPreprocessingBaseClass-fit"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentationPreprocessingBaseClass-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentationPreprocessingBaseClass$fit(X, y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A list of \link[tibble:tibble]{tibble::tibble}s specifying a sample of persistence
diagrams.}

\item{\code{y}}{An integer vector specifying persistence diagram labels (unused for
now).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentationPreprocessingBaseClass-transform"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentationPreprocessingBaseClass-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentationPreprocessingBaseClass$transform(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A list of \link[tibble:tibble]{tibble::tibble}s specifying a sample of persistence
diagrams.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentationPreprocessingBaseClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentationPreprocessingBaseClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentationPreprocessingBaseClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
