% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgudhi-package.R
\docType{package}
\name{rgudhi-package}
\alias{rgudhi}
\alias{rgudhi-package}
\title{rgudhi: An R Interface to GUDHI}
\description{
Provides an R interface to the Python package GUDHI.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LMJL-Alea/rgudhi}
  \item \url{https://lmjl-alea.github.io/rgudhi/}
  \item Report bugs at \url{https://github.com/lmjl-alea/rgudhi/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@math.cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

}
\keyword{internal}
