% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_circle}
\alias{seq_circle}
\title{Circular Sequence Generation}
\usage{
seq_circle(n)
}
\arguments{
\item{n}{An integer value specifying the number of points in the sequence.}
}
\value{
A \link[base:list]{base::list} of length-2 numeric vectors storing 2D points evenly
spaced on the unit circle.
}
\description{
Generates a sequence of 2D points evenly spaced on the unit circle.
}
\examples{
seq_circle(10)
}
