% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorflow-module.R
\docType{data}
\name{tf}
\alias{tf}
\title{Main TensorFlow module}
\format{
TensorFlow module
}
\usage{
tf
}
\description{
Interface to main TensorFlow  module. Provides access to top level classes
and functions as well as sub-modules (e.g. \code{tf$nn},
\code{tf$contrib$learn}, etc.).
}
\examples{
\dontshow{if (reticulate::py_module_available("tensorflow")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
hello <- tf$constant('Hello, TensorFlow!')
zeros <- tf$Variable(tf$zeros(tf$shape(1L)))
tf$print(hello)
tf$print(zeros)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
