% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_ragged_tensor}
\alias{to_ragged_tensor}
\title{Convert Persistence Diagrams to Ragged Tensors}
\usage{
to_ragged_tensor(diagrams)
}
\arguments{
\item{diagrams}{A list of numeric matrices of dimension \eqn{n \times 2}
specifying a sample of \eqn{n} persistence diagrams.}
}
\value{
A ragged tensor of shape \eqn{n \times \mathrm{None} \times 2}.
}
\description{
Convert Persistence Diagrams to Ragged Tensors
}
\examples{
dgm <- list(
 matrix(c(0, 1, 0, 2), ncol = 2, byrow = TRUE),
 matrix(c(0, 1, 0, 2, 0, 3), ncol = 2, byrow = TRUE)
)
dgm
to_ragged_tensor(dgm)
}
