% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-vector-methods.R
\name{VectorRepresentationStep}
\alias{VectorRepresentationStep}
\title{Vector Representation Step}
\description{
Vector Representation Step

Vector Representation Step
}
\author{
Mathieu Carrière
}
\keyword{internal}
\section{Super classes}{
\code{rgudhi::PythonClass} -> \code{rgudhi::SKLearnClass} -> \code{VectorRepresentationStep}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VectorRepresentationStep-apply}{\code{VectorRepresentationStep$apply()}}
\item \href{#method-VectorRepresentationStep-fit}{\code{VectorRepresentationStep$fit()}}
\item \href{#method-VectorRepresentationStep-transform}{\code{VectorRepresentationStep$transform()}}
\item \href{#method-VectorRepresentationStep-fit_transform}{\code{VectorRepresentationStep$fit_transform()}}
\item \href{#method-VectorRepresentationStep-clone}{\code{VectorRepresentationStep$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="get_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-get_python_class'><code>rgudhi::PythonClass$get_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="set_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-set_python_class'><code>rgudhi::PythonClass$set_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="SKLearnClass" data-id="get_params"><a href='../../rgudhi/html/SKLearnClass.html#method-SKLearnClass-get_params'><code>rgudhi::SKLearnClass$get_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="SKLearnClass" data-id="set_params"><a href='../../rgudhi/html/SKLearnClass.html#method-SKLearnClass-set_params'><code>rgudhi::SKLearnClass$set_params()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorRepresentationStep-apply"></a>}}
\if{latex}{\out{\hypertarget{method-VectorRepresentationStep-apply}{}}}
\subsection{Method \code{apply()}}{
Applies the class on a single persistence diagram and
outputs the result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorRepresentationStep$apply(diag)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{diag}}{A 2-column \link[tibble:tibble]{tibble::tibble} specifying a persistence diagram.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link[tibble:tibble]{tibble::tibble} storing the requested vector representation of
the persistence diagram in a table suitable for visualization.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorRepresentationStep-fit"></a>}}
\if{latex}{\out{\hypertarget{method-VectorRepresentationStep-fit}{}}}
\subsection{Method \code{fit()}}{
Fits the class on a sample of persistence diagrams.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorRepresentationStep$fit(X, y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A list of 2-column \link[tibble:tibble]{tibble::tibble}s specifying a sample of
persistence diagrams.}

\item{\code{y}}{An integer vector specifying persistence diagram labels (unused for
now).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The class itself invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorRepresentationStep-transform"></a>}}
\if{latex}{\out{\hypertarget{method-VectorRepresentationStep-transform}{}}}
\subsection{Method \code{transform()}}{
Applies the class on a sample of persistence diagrams.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorRepresentationStep$transform(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A list of 2-column \link[tibble:tibble]{tibble::tibble}s specifying a sample of
persistence diagrams.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of \link[tibble:tibble]{tibble::tibble}s storing the requested vector
representations of the persistence diagrams in a table suitable for
visualization.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorRepresentationStep-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-VectorRepresentationStep-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
Applies sequentially the \verb{$fit()} and \verb{$transform()} methods
on a sample of persistence diagrams in a more efficient way than
calling them directly.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorRepresentationStep$fit_transform(X, y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A list of 2-column \link[tibble:tibble]{tibble::tibble}s specifying a sample of
persistence diagrams.}

\item{\code{y}}{An integer vector specifying persistence diagram labels (unused for
now).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of \link[tibble:tibble]{tibble::tibble}s storing the requested vector
representations of the persistence diagrams in a table suitable for
visualization.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorRepresentationStep-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VectorRepresentationStep-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorRepresentationStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
