% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representations.R
\name{BirthPersistenceTransform}
\alias{BirthPersistenceTransform}
\title{Representation Module - Birth Persistence Transform}
\description{
This is a class for the affine transformation \eqn{(x,y) \mapsto
  (x,y-x)} to be applied on persistence diagrams.
}
\examples{

## ------------------------------------------------
## Method `BirthPersistenceTransform$new`
## ------------------------------------------------

n <- 10
X <- lapply(
  seq(0, 2 * pi, len = n + 1)[1:n],
  function(.x) c(cos(.x), sin(.x))
)
if (reticulate::py_module_available("gudhi")) {
  ac <- AlphaComplex$new(points = X)
  st <- ac$create_simplex_tree()
  dgm <- st$compute_persistence()$persistence_intervals_in_dimension(0)
  bpt <- BirthPersistenceTransform$new()
  bpt$apply(dgm)
  bpt$fit_transform(list(dgm))
}
}
\author{
Mathieu Carrière
}
\section{Super classes}{
\code{rgudhi::PythonClass} -> \code{rgudhi::SKLearnClass} -> \code{\link[rgudhi:RepresentationBaseClass]{rgudhi::RepresentationBaseClass}} -> \code{BirthPersistenceTransform}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BirthPersistenceTransform-new}{\code{BirthPersistenceTransform$new()}}
\item \href{#method-BirthPersistenceTransform-clone}{\code{BirthPersistenceTransform$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="get_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-get_python_class'><code>rgudhi::PythonClass$get_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="PythonClass" data-id="set_python_class"><a href='../../rgudhi/html/PythonClass.html#method-PythonClass-set_python_class'><code>rgudhi::PythonClass$set_python_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="SKLearnClass" data-id="get_params"><a href='../../rgudhi/html/SKLearnClass.html#method-SKLearnClass-get_params'><code>rgudhi::SKLearnClass$get_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="SKLearnClass" data-id="set_params"><a href='../../rgudhi/html/SKLearnClass.html#method-SKLearnClass-set_params'><code>rgudhi::SKLearnClass$set_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="RepresentationBaseClass" data-id="apply"><a href='../../rgudhi/html/RepresentationBaseClass.html#method-RepresentationBaseClass-apply'><code>rgudhi::RepresentationBaseClass$apply()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="RepresentationBaseClass" data-id="fit"><a href='../../rgudhi/html/RepresentationBaseClass.html#method-RepresentationBaseClass-fit'><code>rgudhi::RepresentationBaseClass$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="RepresentationBaseClass" data-id="fit_transform"><a href='../../rgudhi/html/RepresentationBaseClass.html#method-RepresentationBaseClass-fit_transform'><code>rgudhi::RepresentationBaseClass$fit_transform()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rgudhi" data-topic="RepresentationBaseClass" data-id="transform"><a href='../../rgudhi/html/RepresentationBaseClass.html#method-RepresentationBaseClass-transform'><code>rgudhi::RepresentationBaseClass$transform()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BirthPersistenceTransform-new"></a>}}
\if{latex}{\out{\hypertarget{method-BirthPersistenceTransform-new}{}}}
\subsection{Method \code{new()}}{
The \code{\link{BirthPersistenceTransform}} constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BirthPersistenceTransform$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An object of class \code{\link{BirthPersistenceTransform}}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{n <- 10
X <- lapply(
  seq(0, 2 * pi, len = n + 1)[1:n],
  function(.x) c(cos(.x), sin(.x))
)
if (reticulate::py_module_available("gudhi")) {
  ac <- AlphaComplex$new(points = X)
  st <- ac$create_simplex_tree()
  dgm <- st$compute_persistence()$persistence_intervals_in_dimension(0)
  bpt <- BirthPersistenceTransform$new()
  bpt$apply(dgm)
  bpt$fit_transform(list(dgm))
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BirthPersistenceTransform-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BirthPersistenceTransform-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BirthPersistenceTransform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
