% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgudhi-package.R
\docType{package}
\name{rgudhi-package}
\alias{rgudhi}
\alias{rgudhi-package}
\title{rgudhi: An Interface to the GUDHI Library for Topological Data Analysis}
\description{
Provides an interface to the GUDHI library which is a generic open source C++ library, with a Python interface, for topological data analysis (TDA) and higher dimensional geometry understanding. The library offers state-of-the-art data structures and algorithms to construct simplicial complexes and compute persistent homology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LMJL-Alea/rgudhi}
  \item \url{https://lmjl-alea.github.io/rgudhi/}
  \item Report bugs at \url{https://github.com/lmjl-alea/rgudhi/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

Other contributors:
\itemize{
  \item GUDHI Editorial Board (https://gudhi.inria.fr/people/) [contributor]
}

}
\keyword{internal}
